update: muchas cosas

+++ project/src/com/hermes/dao/IGenericDAO.java
@@ -6,7 +6,7 @@ public interface IGenericDAO<T> {
 	
 	public String getNameClass();
 	public List<T> getAll();
-	public T getById(int id);
-	public boolean deteleById(int id);
+	public T getById(Long id);
+	public boolean deteleById(Long id);
 
 }
